# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "B Capital Portfolio Intelligence API",
    "version": "1.0.0",
    "description": "Read API over 136 B Capital portfolio records (135 portfolio companies + the firm) with brand tokens, flagged ownership estimates, news and OnDemand-workflow-scored sentiment (model predefined-deepseek-flash — DeepSeek Flash v4.1). Write-back is a single POST /ingest protected by X-Ingest-Secret.",
    "contact": {
      "name": "B Capital Portfolio Intelligence"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://serverless.on-demand.io/apps/bcap-portfolio-api",
      "description": "Deployed API"
    }
  ],
...
  "paths": {
...
    "/companies": {
      "get": {
        "operationId": "listCompanies",
        "tags": [
          "companies"
        ],
        "summary": "List portfolio companies with filters and pagination",
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "description": "Exact sector, e.g. Technology | Healthcare | Energy & Resilience | Opportunistic",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "description": "Exact region, e.g. North America | Asia | Europe | Middle East",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Substring match on status (private, public, firm, unicorn, renamed…)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "description": "B Capital role",
            "schema": {
              "type": "string",
              "enum": [
                "lead",
                "co-lead",
                "participant",
                "unknown",
                "firm"
              ]
            }
          },
          {
            "name": "focus",
            "in": "query",
            "required": false,
            "description": "true → only the 5 focus companies",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "1-based page",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size (1-200)",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "description": "name | sentiment | updated_at | ticket",
            "schema": {
              "type": "string",
              "default": "name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated companies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Company"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "total_pages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/companies/{slug}": {
      "get": {
        "operationId": "getCompany",
        "tags": [
          "companies"
        ],
        "summary": "One company by slug (full record incl. brand tokens, estimates, news, sentiment)",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Company slug, e.g. perplexity-ai, apptronik, fervo-energy, flutterwave, writer, b-capital"
          }
        ],
        "responses": {
          "200": {
            "description": "Company",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Company"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Unknown slug",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/companies/{slug}/news": {
      "get": {
        "operationId": "getCompanyNews",
        "tags": [
          "companies"
        ],
        "summary": "News items for a company, newest first",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Company slug, e.g. perplexity-ai, apptronik, fervo-energy, flutterwave, writer, b-capital"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Max items",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "description": "news | status_change | social | funding",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "News list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/NewsItem"
                      }
                    },
                    "company": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Unknown slug",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/search": {
      "get": {
        "operationId": "search",
        "tags": [
          "search"
        ],
        "summary": "Full-text search over company name/sector/status/HQ and news titles/summaries",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Query text (min 2 chars)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Max hits per type",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "q": {
                      "type": "string"
                    },
                    "companies": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "news": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/NewsItem"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing q",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ingestSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Ingest-Secret"
      }
    },
    "schemas": {
...
      "NewsItem": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": [
              "string",
              "null"
            ]
          },
          "kind": {
            "type": "string"
          },
          "published_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "sentiment_score": {
            "type": [
              "number",
              "null"
            ]
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "message",
          "errorCode"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      }
    }
  }
}